+++ jobkit/engine/src/main/java/tv/hd3g/jobkit/engine/RunnableWithException.java
@@ -35,4 +35,7 @@ public interface RunnableWithException {
 	static RunnableWithException fromRunnable(final Runnable r) {
 		return r::run;
 	}
+
+	RunnableWithException nothing = () -> {
+	};
 }