+++ jobkit/engine/src/test/java/tv/hd3g/jobkit/engine/SpoolerTest.java
@@ -8,6 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
+import static tv.hd3g.jobkit.engine.RunnableWithException.nothing;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -16,9 +17,6 @@ import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
 class SpoolerTest {
-
-	private static final RunnableWithException emptyRunnable = () -> {
-	};
 
 	@Mock
 	ExecutionEvent event;
@@ -59,30 +57,27 @@ class SpoolerTest {
 
 	@Test
 	void testShutdown() {
-		final var before = spooler.getExecutor("A").addToQueue(emptyRunnable, "empty1", 0, e -> {
+		final var before = spooler.getExecutor("A").addToQueue(nothing, "empty1", 0, e -> {
 		});
 		assertTrue(before);
 
 		spooler.shutdown();
 
-		final var after = spooler.getExecutor("A").addToQueue(emptyRunnable, "empty2", 0, e -> {
-		});
-		assertFalse(after);
-
-		verify(event, Mockito.times(1)).shutdownSpooler(any(Supervisable.class));
-	}
-
-	@Test
-	void testWaitToClose() {
-		spooler.getExecutor("A");
-
-		spooler.waitToClose();
-		final var after = spooler.getExecutor("A").addToQueue(emptyRunnable, "empty", 0, e -> {
+		final var after = spooler.getExecutor("A").addToQueue(nothing, "empty2", 0, e -> {
 		});
 		assertFalse(after);
 
 		assertNull(spooler.getExecutor("B"));
 		verify(event, Mockito.times(1)).shutdownSpooler(any(Supervisable.class));
+
+		spooler.shutdown();
 	}
 
+	@Test
+	void testShutdown_empty() {
+		spooler.shutdown();
+
+		assertNull(spooler.getExecutor("B"));
+		verify(event, Mockito.times(1)).shutdownSpooler(any(Supervisable.class));
+	}
 }