+++ jobkit/springboot-service/src/main/java/tv/hd3g/jobkit/mod/component/JobKitOnBootStartsBackgServices.java
@@ -18,6 +18,7 @@ package tv.hd3g.jobkit.mod.component;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.springframework.beans.factory.DisposableBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.event.EventListener;
@@ -26,7 +27,7 @@ import org.springframework.stereotype.Component;
 import tv.hd3g.jobkit.engine.JobKitEngine;
 
 @Component
-public class JobKitOnBootStartsBackgServices {
+public class JobKitOnBootStartsBackgServices implements DisposableBean {
 	private static Logger log = LogManager.getLogger();
 
 	@Autowired
@@ -36,5 +37,11 @@ public class JobKitOnBootStartsBackgServices {
 	public void startsBackgroundServices() {
 		log.debug("App has booted, start background services...");
 		jobKitEngine.onApplicationReadyRunBackgroundServices();
+	}
+
+	@Override
+	public void destroy() throws Exception {
+		log.debug("App want to close: shutdown jobKitEngine");
+		jobKitEngine.shutdown();
 	}
 }